Move StreamFold and related functionality into StreamFoldFunction
This is part of ensuring that only function related matters are kept in
the hot loop.

For clarity, other matters related to the function are also moved.
Repr information has been stripped out of StreamFold struct, before
moving StreamFold to StreamFoldFunction.
